version 2.0

+++ src/redux/slices/auth-slice.js
@@ -1,6 +1,7 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import { Toast } from "../../components/Toast";
 import { redirect } from "react-router-dom";
+import Swal from "sweetalert2";
 
 
 export const extraLoginAction = createAsyncThunk("authSlice/login", async (data) => {
@@ -26,7 +27,7 @@ const authSlice = createSlice({
             // navigate('/login');
             Toast.fire({
                 icon: 'success',
-                title: 'You Logged Out'
+                title: 'You have logged out successfully'
             })
         }
     },
@@ -35,16 +36,30 @@ const authSlice = createSlice({
         builder
             .addCase(extraLoginAction.fulfilled, (state, action) => {
                 const user = action.payload;
-                // console.log(action, state);
                 // store user details and basic auth data in local storage to keep user logged in between page refreshes
-                localStorage.setItem('user', JSON.stringify(user));
-                state.user = user;
-                Toast.fire({
-                    icon: 'success',
-                    title: 'You Logged In succesfully'
-                })
-                // get return url from location state or default to home page
-                redirect("/");
+                if (user.statusCode === 401) {
+                    Swal.fire({
+                        title: "You don't have Account, Do you want to sign up?",
+                        showDenyButton: true,
+                        confirmButtonText: 'Sign up',
+                        denyButtonText: `return to login`,
+                    }).then((result) => {
+                        if (result.isConfirmed) {
+                            window.location.href = 'http://localhost:3000/signup';
+                        }
+                    })
+
+                } else {
+                    localStorage.setItem('user', JSON.stringify(user));
+                    state.user = user;
+                    Toast.fire({
+                        icon: 'success',
+                        title: 'You have logged in successfully'
+                    })
+                    // get return url from location state or default to home page
+                    redirect("/");
+                }
+
             })
 
     }
